fix:update package.json versions and rename ui files
