+++ package.json
@@ -13,8 +13,8 @@
     "@reduxjs/toolkit": "^2.11.2",
     "axios": "^1.13.4",
     "clsx": "^2.1.1",
-    "next": "^16.1.6",
-    "next-intl": "^4.8.3",
+    "next": "^16.1.3",
+    "next-intl": "^4.8.2",
     "normalize.css": "^8.0.1",
     "pnpm": "^10.30.3",
     "react": "19.2.3",
